+++ apk/Changelog.md
@@ -9,6 +9,14 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Web Version July 13 2022 and Apk Version 3.0.334 and up
+==============
+
+* Update channel on the home screen to work on the new Twitch API
+* Update notificatons to work on the new Twitch API
+* Is now necessary to have added a user and a key for all channel on the home screen
+* General improves
+
 Web Version July 12 2022
 ==============